# reduce impact of memory leaks related to editor