--- a/src/vs/editor/browser/view.ts
+++ b/src/vs/editor/browser/view.ts
@@ -399,6 +399,8 @@ export class View extends ViewEventHandler {
 		}
 
 		super.dispose();
+
+		dom.clearNodeRecursively(this.domNode.domNode);
 	}
 
 	private _scheduleRender(): void {