--- a/src/vs/editor/browser/view.ts
+++ b/src/vs/editor/browser/view.ts
@@ -400,6 +400,12 @@ export class View extends ViewEventHandler {
 
 		super.dispose();
 
+		// See https://github.com/microsoft/vscode/pull/219297
+		// Reduces the impact of detached dom node memory leaks
+		// E.g when a memory leak occurs in a child component
+		// of the editor (e.g. Minimap, GlyphMarginWidgets,...)
+		// the editor dom (linked to the child dom) should not be
+		// leaked as well.
 		dom.clearNodeRecursively(this.domNode.domNode);
 	}