@SimonSiefke 2024-07-01
Thanks for the review! I added a comment with a link to this pull request. And also updated the comment above to include the script used for measuring canvas counts and detached dom nodes.